Code quality: Refactor BatteryIcon and bump H.NotifyIcon.WinUI from 2.0.99 to 2.0.104
